Split out the legacy system into a seperate area of the project for the moment.
