+++ app/page.tsx
@@ -1,22 +1,5 @@
-//@ts-expect-error
-import { getRequestContext } from "@cloudflare/next-on-pages";
-import { drizzle } from "drizzle-orm/d1";
-
-import { document } from "./schema";
+import { redirect } from "next/navigation";
 
 export default async function Page() {
-  const { env } = getRequestContext();
-  const db = drizzle(env.MAIN_DB as D1Database);
-  const result = await db.select().from(document).all();
-
-  return (
-    <div>
-      {result.map((item) => (
-        <div key={item.id}>
-          <p>ID: {item.id}</p>
-          <p>Name: {item.name}</p>
-        </div>
-      ))}
-    </div>
-  );
+  redirect("/2024");
 }
